refactor(server): split plan-runs.test.ts into five focused test files alongside plan-runs.test-helpers.ts (warren-64f6)
