+++ .mulch/expertise/server.jsonl
@@ -82,3 +82,4 @@
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-27T22:48:05.083Z","evidence":{"commit":"3cfbdbb751409900a4ca73ba873071a9a82b13ed"},"relates_to":["mx-5009ad"],"content":"src/server/main.ts (warren-8d3d / pl-9088 step 10) is split into src/server/main/ subdirectory: utils.ts (defaultSpawn + closeDatabase + redactDbUrl + resolvePgPoolMax + parseTrueEnv + parseIntEnv), logging.ts (6 pino→narrow logger adapters: bridge/probe/scheduler/planRun/pause/previewEviction), preview-wiring.ts (createPreviewAuthAndProxy — R-19/SPEC §11.L token × mode × host matrix), deps.ts (buildServerDeps — ServerDeps literal + plotAggregator), index.ts (bootServer orchestrator + CLI entry + public type re-exports, 472 lines). Test files split 1:1 alongside (utils.test.ts/logging.test.ts/preview-wiring.test.ts + the original resolvePgPoolMax suite at main/index.test.ts) so the extracted helpers regain direct coverage — otherwise the split would have dropped lines coverage below the 90% floor since boot-only code lives in the 0%-executed bootServer body. Importers (src/cli/commands/serve.{ts,test.ts}, src/server/index.ts) re-point to main/index.ts. scripts/file-size-budgets.json entry for main.ts removed; biome.json noExcessiveCognitiveComplexity override updated main.ts -> main/index.ts (bootServer body still trips the threshold). The CLI entry shebang (import.meta.main block) lives in main/index.ts; bun run src/server/main/index.ts replaces the prior src/server/main.ts invocation in the Phase-12 supervisor exec path.","id":"mx-131efe"}
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-28T03:05:28.261Z","evidence":{"commit":"245190a440fa18c003d43464dcb4f4a1cce127e7"},"dir_anchors":["src/server/handlers"],"content":"plots/ index.ts is a composing route table for plots subdirectory, re-exporting the split handlers list, sync, workbench and remaining monolithic plots.ts handlers.","id":"mx-b7697e"}
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-28T03:29:35.665Z","evidence":{"commit":"bf1bbc8819f28805972385f605bcb3512c92315b"},"dir_anchors":["src/server/handlers"],"content":"Split src/server/handlers/runs.ts into runs/ subdirectory domain modules (dispatch, lifecycle, events, preview, pause-resume) and runs.test.ts into modular test files (runs.dispatch.test, runs.interactive.test, runs.events.test) with shared test helpers (warren-6566)","id":"mx-cad464"}
+{"type":"convention","classification":"tactical","recorded_at":"2026-05-28T03:37:50.242Z","evidence":{"seeds":"warren-64f6","commit":"709ddbf75aede9ca81ee9ae40d4f60f7253f6fc3"},"content":"Split src/server/handlers/plan-runs.test.ts (1318 lines) into plan-runs.test-helpers.ts and five focused test files under 500 lines: create, plot, get, cancel, events. Removed old monolithic test file and updated file-size budget.","id":"mx-04acb7"}
